EVG-2134: add panic handlers to all agent threads
