+++ command/s3_copy.go
@@ -80,8 +80,6 @@ func (c *s3copy) ParseParams(params map[string]interface{}) error {
 // validateParams is a helper function that ensures all
 // the fields necessary for carrying out an S3 copy operation are present
 func (c *s3copy) validateParams() error {
-	var err error
-
 	if c.AwsKey == "" {
 		return errors.New("s3 AWS key cannot be blank")
 	}
