+++ config.go
@@ -4,12 +4,9 @@ import (
 	"io/ioutil"
 	"time"
 
-	"github.com/evergreen-ci/evergreen/util"
 	"github.com/mongodb/grip"
-	"github.com/mongodb/grip/message"
 	"github.com/mongodb/grip/send"
 	"github.com/pkg/errors"
-	"golang.org/x/net/context"
 	"gopkg.in/yaml.v2"
 )
 
@@ -354,32 +351,6 @@ func (s *Settings) GetSender() (send.Sender, error) {
 	}
 
 	return send.NewConfiguredMultiSender(senders...), nil
-}
-
-// SystemInfoCollector is meant to run in a goroutine and log
-// aggregate system resource utilization (cpu, memory, network, i/o)
-// every 15 seconds. The information is logged to process' default
-// grip logger.
-//
-// In general, the collector should run in the background of the API
-// server and the UI server.
-func SystemInfoCollector(ctx context.Context) {
-	defer util.RecoverLogStackTraceAndContinue("system info collector")
-	const sysInfoLoggingInterval = 15 * time.Second
-	timer := time.NewTimer(0)
-	defer timer.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			grip.Info("system logging operation canceled")
-			return
-		case <-timer.C:
-			grip.Info(message.CollectSystemInfo())
-			grip.Info(message.CollectGoStats())
-			timer.Reset(sysInfoLoggingInterval)
-		}
-	}
 }
 
 // ConfigValidator is a type of function that checks the settings
